edu modules; eg decay=release canvas
