+++ index.html
@@ -11,6 +11,7 @@
 <body>
     <div id="nav">
     </div>
+    <div id="eduModal"></div>
     <div id="editModal">
     </div>
     <div id="saveModal">
